Safari-611.1.10/LayoutTests/imported/blink/fast/events/eventlisteneroptions/capture_query.html

<!DOCTYPE HTML>
<body/>
<script src="../../../../../resources/testharness.js"></script>
<script src="../../../../../resources/testharnessreport.js"></script>

<script>
test(function(t) {
    var supportsCapture = false;
    var query_function = function(e) {};
    var query_options = {
        get capture() {
            supportsCapture = true;
            return false;
        },
        get dummy() {
            assert_unreached("dummy value getter invoked");
            return false;
        }
    };

    document.addEventListener('test_event', query_function, query_options);
    assert_true(supportsCapture);
    supportsCapture = false;
    document.removeEventListener('test_event', query_function, query_options);
    assert_true(supportsCapture);
    t.done();
}, "Supports Capture");
</script>
